+++ README.md
@@ -1,6 +1,6 @@
 # React dynamic list of Posts
 - Replace `<your_account>` with your Github username in the
-  [DEMO LINK](https://CloudSweeT.github.io/react_dynamic-list-of-posts/)
+  [DEMO LINK](https://CouldSweeT.github.io/react_dynamic-list-of-posts/)
 - Follow the [React task guideline](https://github.com/mate-academy/react_task-guideline#react-tasks-guideline)
 - Use [React TypeScript cheat sheet](https://mate-academy.github.io/fe-program/js/extra/react-typescript)
