+++ src/App.tsx
@@ -1,12 +1,18 @@
-import React, { useState } from 'react';
+import React, { useState, useEffect } from 'react';
 import './App.scss';
 import './styles/general.scss';
 import { PostsList } from './components/PostsList';
 import { PostDetails } from './components/PostDetails';
+import { getUsers } from './api/post';
 
 const App: React.FC = () => {
   const [selectedUserId, handleUserId] = useState(0);
   const [selectedPostId, handlePostId] = useState<number | null>(null);
+  const [users, setUsers] = useState<User[]>([]);
+
+  useEffect(() => {
+    getUsers().then(result => setUsers(result));
+  }, []);
 
   return (
     <div className="App">
@@ -15,23 +21,20 @@ const App: React.FC = () => {
           Select a user: &nbsp;
 
           <select
-            className="App__user-selector"
             id="select"
-            name="select"
+            className="App__user-selector"
             value={selectedUserId}
-            onChange={event => handleUserId(+event.currentTarget.value)}
+            onChange={event => handleUserId(+event.target.value)}
           >
             <option value="0">All users</option>
-            <option value="1">Leanne Graham</option>
-            <option value="2">Ervin Howell</option>
-            <option value="3">Clementine Bauch</option>
-            <option value="4">Patricia Lebsack</option>
-            <option value="5">Chelsey Dietrich</option>
-            <option value="6">Mrs. Dennis Schulist</option>
-            <option value="7">Kurtis Weissnat</option>
-            <option value="8">Nicholas Runolfsdottir V</option>
-            <option value="9">Glenna Reichert</option>
-            <option value="10">Leanne Graham</option>
+            {users && (
+              users.map((user : User) => {
+
+                return (
+                  <option key={user.id} value={user.id}>{user.name}</option>
+                );
+              })
+            )}
           </select>
         </label>
       </header>
